Level 3
Set the scene for level 3, with another shooting alien, this time he is
shooting a fireball. Set the power up planet to give the player extra
time when collided with, and set the player to shoot starts. This time,
the obstacle is a rocket which can cause damage on collision with
player